Implemented storing temporal restocking plans
Each VendingMachine now knows how often it should be restocked, while
its current VMLayout knows when it will be next be restocked.  This
separation exists because a change to the restocking interval that
postpones the next stocking visit could potentially result in expired
products' remaining in the machine.  Thus, it is possible for the
current layout's next stocking date to be temporarily out of sync with
its machine's defined schedule.