Started working on CustomerLoginScreenGUI. This commit should get squashed later.
Got the customer login screen to display decently well. This commit should be squashed.

Need to add status bar and move some code.

Added BaseGUI, StatusBar, and CustomerLoginScreenGUI classes.

The CustomerLoginScreenGUI class builds the GUI for the
CustomerLoginScreen. It currently has no functionality.

Added BaseGUI, which will act as a base for all GUIs needed for this
project. It will include the status bar and maintain a standard size. It
might also need some refactoring later.

Added StatusBar class, which is the implementation of status bar.
Currently it doesn't look very nice, but includes a label and a
background color.