Changed BaseGUI from an abstract class that GUI screens would have extended to a container in which GUI screens will be placed.
Removed abstract addComponents() method from BaseGUI.

Removed setProperBorder() static method from BaseGUI because the new
setContentPanel() method handles what setProperBorder used to be used
for.

Added setContentPanel() method.
	This method is used to swap the content that is currently being
	displayed. This is useful because it allows the programmer to change
	the screen (e.g. changing the customer login screen to the customer
	purchase screen) without the status bar or the window changing.

Changed addBaseComponents() to be private because it should only be
called (once) from the constructor.

Modified CustomerLoginScreenGUI so that it works with these changes.
	Changed it from extending BaseGUI to JPanel.

	Modified the constructor to handle the change in super class.
	The constructor now calls addComponents().

	Removed mainPanel local variable in addComponents() because it is
	no longer need (it is replaced by the instance of
	CustomerLoginScreenGUI).

Changed the main method in CustomerLoginScreenGUI to handle work with
the changes.