Chaned the ConditionButton to use ConditionButtonConditions to check conditions instead of being an abstract class.
Made ConditonButton a concrete class.

Added ConditionButtonCondition interface. It includes a checkCondition()
method that is used by ConditionButton to determine if the button should
be enabled or not.

Added an addCondition() method to ConditionButton that takes a
ConditionButtonCondition and adds it to a list of condition to be
checked.

checkAndSetEnabled() in ConditionButton now checks a list of conditions
and enables the button iff all of the conditions are true.